+++ src/main/groovy/ru/vyarus/gradle/plugin/animalsniffer/report/AnimalSnifferReports.groovy
@@ -16,10 +16,36 @@ import org.gradle.api.tasks.Internal
 interface AnimalSnifferReports extends ReportContainer<SingleFileReport> {
 
     /**
+     * IMPORTANT: this report is required in order to build text report and console output (because animalsniffer
+     * itself could run in another jvm).
+     * <p>
+     * Report columns:
+     * <ul>
+     *     <li>Signature name
+     *     <li>Source file
+     *     <li>Line number
+     *     <li>Field (when problem in field)
+     *     <li>Error message (when message parse failed would contain complete (unparsed) error message)
+     *     <li>True when failed to parse animalsniffer error
+     * </ul>
+     *
+     * @return internal csv report used for console output
+     */
+    @Internal
+    SingleFileReport getCsv()
+
+    /**
      * @return Animal-sniffer text report
      */
     @Internal
     SingleFileReport getText()
+
+    /**
+     * Configures the csv report.
+     *
+     * @param action The Configuration closure/action.
+     */
+    void csv(Action<SingleFileReport> action);
 
     /**
      * Configures the text report.