+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/ErrorFormatTest.groovy
@@ -87,16 +87,16 @@ class ErrorFormatTest extends Specification {
 
         when: "line with field"
         ReportMessage msg = FormatUtils.parse(
-                '/opt/foo/invalid/Sample.java Field \'field\': Undefined reference: java.nio.file.Path'
+                '/opt/foo/invalid/Sample.java Field field: Undefined reference: java.nio.file.Path'
                 , roots)
         then: "parsed"
         msg.source == 'invalid.Sample.java'
         msg.line == null
-        msg.field == 'field \'field\''
+        msg.field == 'field'
         msg.code == 'java.nio.file.Path'
 
         then: "formatting"
-        FormatUtils.formatForFile(msg, false) == "invalid.Sample:1 (field 'field')  Undefined reference: java.nio.file.Path"
-        FormatUtils.formatForConsole(msg, false).replaceAll('\r', '') == "[Undefined reference] invalid.(Sample.java:1) field 'field'\n  >> java.nio.file.Path\n"
+        FormatUtils.formatForFile(msg, false) == "invalid.Sample:1 (#field)  Undefined reference: java.nio.file.Path"
+        FormatUtils.formatForConsole(msg, false).replaceAll('\r', '') == "[Undefined reference] invalid.(Sample.java:1) #field\n  >> java.nio.file.Path\n"
     }
 }